% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NISloadR-package.R
\docType{package}
\name{NISloadR-package}
\alias{NISloadR}
\alias{NISloadR-package}
\title{NISloadR: Load National Inpatient Sample (NIS) Data into R}
\description{
A package to directly load National Inpatient Sample (NIS) data into R.
}
